chore(rhomb): build script updated to remove download directory
Build script updated to remove download directory with "distclean" command.
We were facing issue of getting old package due to same .tgz file names in
download directory. This change will make sure that after distclean we always
get new package. In addition to that README is also updated for typo in
username which was still showing "odroid" instead of "rhomb"

Signed-off-by: Android Linux Media <android.linuxmedia@gmail.com>