rhomb(chore): build script updated to get modified ubuntu for better perfomance
This commit modify build script to get modified ubuntu version for better
performance. Previous ubuntu version was using GDM as display manager which was
consuming CPU and system performance was degraded. In this ubuntu we are using
lightdm instead GDM for better performance. In addition to that we have also
increase CPU frequency in kernel for better performance.

Signed-off-by: Android Linux Media <android.linuxmedia@gmail.com>